+++ src/FileHandlerSE.cpp
@@ -18,7 +18,7 @@ FileHandlerSE::FileHandlerSE(unordered_map<string,string>& barcode_file_map, str
 void FileHandlerSE::init_all_file_handles(double total_buffer_gb){
 	string barcode, sample_name;
 	size_t given_buffer_size_per_file_bytes = int(floor(((total_buffer_gb/(Barcode_file_map->size()*2))*pow(1024,3))));
-	size_t buffer_size_bytes = max(2ul, given_buffer_size_per_file_bytes);
+	size_t buffer_size_bytes = max(size_t(2), given_buffer_size_per_file_bytes);
 
 	for(auto it = Barcode_file_map->begin(); it != Barcode_file_map->end(); it++){
 		barcode = it->first;
